requirements.txt
numpy>=1.24,<3
pandas>=2.0,<3
scikit-learn>=1.3,<2
pyyaml>=6.0
mlflow>=2.9,<3
fastapi>=0.109,<1
uvicorn[standard]>=0.27,<1
prometheus-client>=0.19,<1
